Delete redundant Threads.kt Move Template functions to TemplateUtil.kt
