+++ src/main/kotlin/com/github/kadehar/Main.kt
@@ -1,14 +1,14 @@
 package com.github.kadehar
 
-import com.github.kadehar.async.collectDataAsync
-import com.github.kadehar.async.collectPhrasesAsync
-import com.github.kadehar.async.parseConfigAsync
-import com.github.kadehar.async.parseSummaryAsync
 import com.github.kadehar.client.TelegramApi
 import com.github.kadehar.config.Config
+import com.github.kadehar.config.parseConfig
 import com.github.kadehar.model.Summary
 import com.github.kadehar.template.Message
 import com.github.kadehar.utils.collectTemplateData
+import com.github.kadehar.utils.prepareTemplateData
+import com.github.kadehar.utils.templatePhrases
+import com.github.kadehar.utils.toSummary
 import kotlinx.coroutines.runBlocking
 import retrofit2.Call
 import retrofit2.Response
@@ -16,16 +16,16 @@ import retrofit2.Response
 fun main() {
     var config: Config
     runBlocking {
-        config = parseConfigAsync()
+        config = parseConfig()
     }
 
     var summary: Summary
     var phrases: Map<String, String>
     var data: Map<String, Any>
     runBlocking {
-        summary = parseSummaryAsync(config.project.file)
-        phrases = collectPhrasesAsync(config.project.language)
-        data = collectDataAsync(config.project, summary)
+        summary = toSummary(config.project.file)
+        phrases = templatePhrases(config.project.language)
+        data = prepareTemplateData(config.project, summary)
     }
 
     val map = collectTemplateData(phrases, data)
